requirements-dev.txt
tensorflow
keras
python-opencv
Pillow
imutils